fix: mypage path
