+++ frontend/app/stores/loanAtom.ts
@@ -1,6 +1,5 @@
 import { atom } from 'jotai';
 import { CartProps } from './cartAtom';
-import { atomWithStorage, createJSONStorage } from 'jotai/utils';
 
 // 選択された貸出情報を管理するAtom
 export const selectedLoanAtom = atom<CartProps[]>([]);
